+++ test/stubs/discord-stub.js
@@ -3,14 +3,14 @@ import events from 'events';
 import sinon from 'sinon';
 import discord from 'discord.js';
 
-export default function createDiscordStub(sendStub, guild, discordUsers) {
+export default function createDiscordStub(sendStub, discordUsers) {
   return class DiscordStub extends events.EventEmitter {
     constructor() {
       super();
       this.user = {
         id: 'testid'
       };
-      this.channels = this.guildChannels();
+      this.channels = new discord.Collection();
       this.options = {
         http: {
           cdn: ''
@@ -18,18 +18,49 @@ export default function createDiscordStub(sendStub, guild, discordUsers) {
       };
 
       this.users = discordUsers;
+      this.guilds = new discord.Collection();
+      this.guild = this.createGuildStub();
+      this.guilds.set(this.guild.id, this.guild);
     }
 
-    guildChannels() {
-      const channels = new discord.Collection();
-      channels.set('1234', {
-        name: 'discord',
-        id: '1234',
-        type: 'text',
-        send: sendStub,
-        guild
-      });
-      return channels;
+    addTextChannel(guild, textChannel) {
+      const textChannelData = Object.assign({
+        type: 'text'
+      }, textChannel);
+      const textChannelObj = new discord.TextChannel(guild, textChannelData);
+      textChannelObj.send = sendStub;
+      const permissions = new discord.Collection();
+      textChannelObj.setPermissionStub = (user, perms) => permissions.set(user, perms);
+      textChannelObj.permissionsFor = user => permissions.get(user);
+      this.channels.set(textChannelObj.id, textChannelObj);
+      return textChannelObj;
+    }
+
+    createGuildStub(guildData = {}) {
+      const guild = {
+        id: '1',
+        client: this,
+        roles: new discord.Collection(),
+        members: new discord.Collection(),
+        emojis: new discord.Collection(),
+        channels: new discord.Collection(),
+        addTextChannel: (textChannel) => {
+          const textChannelObj = this.addTextChannel(guild, textChannel);
+          textChannelObj.guild.channels.set(textChannelObj.id, textChannelObj);
+          return textChannelObj;
+        }
+      };
+      Object.assign(guild, guildData);
+      this.guilds.set(guild.id, guild);
+
+      if (guild.id === '1') {
+        guild.addTextChannel({
+          name: 'discord',
+          id: '1234',
+        });
+      }
+
+      return guild;
     }
 
     login() {